Fix conersion warnings

+++ node/silkworm/stagedsync/stage_senders/recovery_farm.hpp
@@ -119,8 +119,8 @@ class RecoveryFarm : public Stoppable {
 
     /* Canonical blocks + headers */
     struct HeaderInfo {
-        HeaderInfo(uint32_t count, const evmc::bytes32& hash) : txn_count(count), block_hash{hash} {};
-        uint32_t txn_count;
+        HeaderInfo(uint64_t count, const evmc::bytes32& hash) : txn_count(count), block_hash{hash} {};
+        uint64_t txn_count;
         evmc::bytes32 block_hash;
     };
     std::vector<HeaderInfo> headers_{};               // Collected canonical headers
